Address PR #12892 review feedback
- Replace Feature:KubeVirt-KIND with RequiresRealKubeVirt requirement label
  so KIND lane includes Feature:KubeVirt and excludes RequiresRealKubeVirt
  instead of double-tagging tests with two features (Casey #1).
- Add back isKINDCluster guards with explicit failure messages explaining
  why the test cannot run on KIND (Casey #2).
- Split live_migration_kind.go: move helpers (kindBIRDPeer, setupKindEBGPPeering,
  expectMigrationFailed) into utils.go, keep only test suite in the file (Casey #3).
- Use existing WithExternalNode() label instead of RequiresExternalNode()
  to avoid duplicate labels with same semantics (skoryk #7).
- Move WithSerial() from describe block to eBGP It only — other tests
  are namespace-scoped and parallel-safe (skoryk #8).
- Add nil guard for net.ParseIP().To4() in querySnapshot and
  queryTORSnapshot to prevent panic on malformed/IPv6 input (skoryk #9).
- Remove duplicate NetworkPolicy test from live_migration.go — the KIND
  ICMP version provides equivalent coverage without requiring a real
  guest OS (skoryk #10).
- Randomize BIRD container name with GenerateRandomName to prevent
  collisions on parallel runs (skoryk #11).
- Change queryWorkerMetric to return (int, error) so failures surface
  the real cause instead of hiding it behind a -1 sentinel (skoryk #13).
- Add non-go:/e2e/pkg/ to CHANGE_IN macro so test file changes trigger
  the KubeVirt KIND CI lane (skoryk #14).
