+++ .semaphore/semaphore.yml
@@ -382,6 +382,7 @@ blocks:
         '/confd/pkg/run/*.go',
         '/crypto/pkg/tls/*.go',
         '/felix/aws/*.go',
+        '/felix/aws/ec2query/*.go',
         '/felix/bpf-apache',
         '/felix/bpf-gpl',
         '/felix/bpf/*.go',
@@ -764,6 +765,7 @@ blocks:
         '/docker/calico/',
         '/felix/',
         '/felix/aws/*.go',
+        '/felix/aws/ec2query/*.go',
         '/felix/bpf/*.go',
         '/felix/bpf/allowsources/*.go',
         '/felix/bpf/arp/*.go',
@@ -1162,6 +1164,7 @@ blocks:
         '/crypto/pkg/tls/*.go',
         '/felix/',
         '/felix/aws/*.go',
+        '/felix/aws/ec2query/*.go',
         '/felix/bpf-apache',
         '/felix/bpf-gpl',
         '/felix/bpf/*.go',
@@ -1860,6 +1863,7 @@ blocks:
         '/confd/pkg/run/*.go',
         '/crypto/pkg/tls/*.go',
         '/felix/aws/*.go',
+        '/felix/aws/ec2query/*.go',
         '/felix/bpf/*.go',
         '/felix/bpf/allowsources/*.go',
         '/felix/bpf/arp/*.go',
@@ -2641,6 +2645,7 @@ blocks:
         '/crypto/pkg/tls/*.go',
         '/e2e/**',
         '/felix/aws/*.go',
+        '/felix/aws/ec2query/*.go',
         '/felix/bpf-apache',
         '/felix/bpf-gpl',
         '/felix/bpf/*.go',
@@ -4377,6 +4382,7 @@ blocks:
         '/crypto/pkg/tls/*.go',
         '/e2e/cmd/k8s/*.go',
         '/e2e/images/rapidclient/*.go',
+        '/e2e/pkg/',
         '/felix/**',
         '/felix/bpf-apache',
         '/felix/bpf-gpl',
@@ -4656,6 +4662,7 @@ blocks:
         '/confd/pkg/run/*.go',
         '/crypto/pkg/tls/*.go',
         '/felix/aws/*.go',
+        '/felix/aws/ec2query/*.go',
         '/felix/bpf-apache',
         '/felix/bpf-gpl',
         '/felix/bpf/*.go',
@@ -4933,6 +4940,7 @@ blocks:
         '/confd/pkg/run/*.go',
         '/crypto/pkg/tls/*.go',
         '/felix/aws/*.go',
+        '/felix/aws/ec2query/*.go',
         '/felix/bpf-apache',
         '/felix/bpf-gpl',
         '/felix/bpf/*.go',
@@ -5276,6 +5284,7 @@ blocks:
         '/confd/pkg/run/*.go',
         '/crypto/pkg/tls/*.go',
         '/felix/aws/*.go',
+        '/felix/aws/ec2query/*.go',
         '/felix/bpf-apache',
         '/felix/bpf-gpl',
         '/felix/bpf/*.go',
